+++ src/zedwallet/Tools.cpp
@@ -371,3 +371,41 @@ bool shutdown(std::shared_ptr<WalletInfo> walletInfo, CryptoNote::INode &node,
     
     return true;
 }
+
+std::vector<std::string> split(const std::string& str, char delim = ' ')
+{
+    std::vector<std::string> cont;
+    std::stringstream ss(str);
+    std::string token;
+    while (std::getline(ss, token, delim)) {
+        cont.push_back(token);
+    }
+    return cont;
+}
+
+bool parseDaemonAddressFromString(std::string& host, int& port, const std::string& address)
+{
+    std::vector<std::string> parts = split(address, ':');
+
+    if (parts.empty())
+    {
+        return false;
+    }
+    else if (parts.size() >= 2)
+    {
+        try
+        {
+            host = parts.at(0);
+            port = std::stoi(parts.at(1));
+            return true;
+        }
+        catch (const std::invalid_argument&)
+        {
+          return false;
+        }
+    }
+
+    host = parts.at(0);
+    port = CryptoNote::RPC_DEFAULT_PORT;
+    return true;
+}
